[ new ] OK by Construction Proof Search for IMLL
This is the way I was describing this work a few weeks ago in
Edinburgh: Resource aware contexts and proof search for IMLL

Starting from the observation that naïve proof search for IMLL is
intractable, we will see how building a notion of resource availability
into the context and massaging the sequent calculus intoa more general
one handling both resource consumption and a notion of "leftovers" of a
subproof allows for a well-structured, well-typed by construction proof
search mechanism.
The fragment we will discuss will consist of `tensor` and `with`.