[ new ] Eval of an untyped, recursive language
The applicative interface provided by delay.agda does not correspond
to the one provided by the delay monad: rather than adding delays, we
max them (cf. the definition of `app`) thus measuring the *depth* of
the recursion rather than the size of the tree.

As a consequence, we don't have to go fetch the result of running `n!`
at a depth higher than `n+1`. Fun.