[ pocpoc ] Representing Cyclic Structures
Using a non-standard semantics for `Desc`riptions, we can
have a universe of finitely represented cyclic structures.

We can equip them with a semantics consisting of infinitely
unrolling the cycle thus reducing proving the correctness of
functions transforming them to proving bisimilarity on the
unfolded versions.

Next step: the definition of an iteration / inductive principle.