[ fix + more ] Adding Omega to the levels
When infering the type for `(t : A)`, we used to need to come up
with a level `l` at which to check that `A` is indeed a type. But
in Type Theory, the traditional approach is to simply have a
`G |- A TYPE` judgment. Extending the levels hierarchy with omega
effectively does that.