[ new ] Bayesian Monad
Inspired by the small calculus presented in Borgström, Gordon,
Greenberg, Margetson and van Gael's Measure Transformer Semantics
for Bayesian Machine Learning, I defined a small `Bayesian` monad
and equipped it with a couple of primitives corresponding to their
`bernouilli` and `observe` constructs.

Their let-bindings are really there to make a distinction between
computations and value usage. This is modelled using do-notations.

We can then run simulations using `runBayesian`: only the valid
runs are retained (i.e. if an `observe` fails, the corresponding
run is thrown away) and the final distribution is normalised thanks
to a cute little circular program.

`checkBayesian` lets us specify the outcome we are interested in.
In the espitemiology example for instance, we want to know the
probability that the patient has the disease so we'd run:

> checkBayesian True epidemiology

and we'd get back:

> [(25 % 322,True)]

where 25 % 322 ~ 0.0776 as stated in the paper.