[ new ] Arbitrary Patterns in let bindings
Rather than having a let binding for pairs only, we define a more
general one able to model more traditional let-bindings as well as
deconstructing arbitrarily nested pairs.

The syntax of patterns requires the user to state explicitly which
variables are going to be used and which ones are of no interest
in the derivation. This is yet another example of the fact that
this approach quotients over the place where leftover resources
are discarded.