[ fix ] Cleaner semantics for `─
The idea is that variables marked as `─ in a pattern are discarded.
By associating an empty Holey ([]) to them rather than a singleton
of the appropriate type, we explicitly throw them away.

The nice side effect is that the deBruijn indices used in the body
of a let-binding are now more intuitive: `z` refers to the first
relevant variable.