[ new ] Semantics of the imperative DSL
It would be nicer to be able to defer the choice of a monad stack.
At the moment the evaluation function is too tied to that choice
which makes many extensions quite hard (e.g. if we were to add a
"print" primitive to the language, we'd need to add a writer monad
somewhere on the stack).