[ new ] Poly-Kinded version of Syntax With Binding
Before this commit we had to use * has the kind of scoped variables.
We would typically work with `Maybe^n Void` as a representation of
a scope with `n` variables.

Now that we have a poly-kinded definition, we are able to use other
approaches. A bunch of them are defined in `Scopes`: `Fin` but also
`Integer` wrapped in a `DeBruijn` newtype tracking at the type level
the number of variables in scope without incuring any runtime heavy
representation (unlike `Fin` whose size is at least linear in the
number represented).