[ refactor ] toStream's type + Show instance
Given that we always use toStream on `CL e 'Zero`, we might as well
define it for `CList e`.

Added a Show instance for `CList` to be able to have a look at CLists
without having to unfold them into streams. Had to use an indirection
via `Apply` because the generic `Show` instance for `Fix ...` only
works for *-based functors. `Show1` is not general enough for our
purpose.