[ new ] map to adjust the exception type
In order to use an `int exn` and a `float exn` together, you need
to wrap them in e.g. `(int, float) sum exn`. This module makes it
possible to write the two embeddings.

Caveat: you need to catch the exception to adjust it so you should
try not to do this too often (not sure whether it'd be possible to
use polymorphic variants to avoid the need for retagging).