[ stuck ] looking for a 1st order repr of Env
Looking for a 1st order representation of environments with a
built-in notion of weakening which can also be used to represent
thinnings.

It's not clear how to combine two of these for disjoint support.
The solution may be to go back to OPEs on one side and environments
on the other.